[SERVER] [LIBRARY] moved missing resources and fixed CAMP tests moved in 7007c65
