+++ brooklyn-library/software/webapp/src/test/resources/java-web-app-and-db-with-policy.yaml
