+++ brooklyn-library/software/webapp/src/test/resources/test-app-with-enrichers-slightly-simpler.yaml
