+++ brooklyn-library/software/webapp/src/test/resources/test-webapp-with-averaging-enricher.yaml
